refactor and add a file 1. add an error-checking feature for cases where the required number count exceeds the range of available numbers 2. separate the config.py file into conf.py and const.py
